refactor(dht): remove BucketIter & create packed_node.rs
