+++ src/toxcore/dht_new/mod.rs
@@ -26,3 +26,4 @@
 pub mod packet;
 pub mod binary_io;
 pub mod kbucket;
+pub mod packed_node;
